Add calculator and render testing
